1. move the config patch's proxy match to the top level , this is to reduce the calling of proxyMatch times,

2. merging envoyfilterWrappers in getting procedure from push context

3. The performance is improved.

Before: 

```
$ go test  -bench=. -benchtime 30s
goos: linux
goarch: amd64
pkg: istio.io/istio/pilot/pkg/networking/core/v1alpha3/envoyfilter
BenchmarkTelemetryV2Filters-4   	  176700	    208609 ns/op
```

With this pr
```
$ go test  -bench=. -benchtime 30s
goos: linux
goarch: amd64
pkg: istio.io/istio/pilot/pkg/networking/core/v1alpha3/envoyfilter
BenchmarkTelemetryV2Filters-4   	  415954	     87968 ns/op

```