@istio-testing 2019-12-25
@hzxuzhonghu: The following test **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
lint_istio | a645a34f7cb4a8bd9d3d39f388260f1a56c3cc08 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/19786/lint_istio/5400) | `/test lint_istio`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->