@istio-testing 2019-12-25
@hzxuzhonghu: The following test **failed**, say `/retest` to rerun them all:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
lint_istio | 1e01353503b38f3e1e7cca4679f9a245a9c94e05 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/19786/lint_istio/5401) | `/test lint_istio`



<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->