@howardjohn 2019-12-26
I do see CPU usage go from 600m to 500m with this patch on a cluster