@howardjohn 2019-12-26
lgtm if we can explain why its now faster, and fix the extra `return` in the test (unless there is a reason for it?). Thanks!