@hzxuzhonghu 2019-12-27
Thanks for the e2e performance test.  And the perfomrance improvement is related to `Reducing the proxyMatch calling times a lot`