@howardjohn 2019-12-27
I don't think proxy match is that cpu intensive though? I would be interested to see profiles before and after this change. I can run it myself though