@howardjohn 2019-12-27
Ah so the main performance benefit is actually merging all the patches. It makes this get called only one time: https://github.com/istio/istio/blob/0a5be9d306e88a5d243dfea0b4400a7460a24bc6/pilot/pkg/networking/core/v1alpha3/envoyfilter/listener_patch.go#L311